pyproject.toml
[tool.poetry]
name = "jobnotify"
version = "0.1.0"
description = ""
authors = ["levch <levchenko.d.a1998@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^2.20"
Telethon = "^1.24.0"
motor = "^3.0.0"
python-i18n = {extras = ["YAML"], version = "^0.3.9"}
openpyxl = "^3.0.10"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
